requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
dj-database-url==3.0.1
Django==4.2.23
django-allauth==65.11.0
django-ckeditor-5==0.2.18
django-js-asset==3.1.2
django-widget-tweaks==1.5.0
gunicorn==23.0.0
idna==3.10
numpy==2.3.2
packaging==25.0
pandas==2.3.1
phpserialize==1.3
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.9.0
